+++ tests/integration/dashboard/scheduled-interactions-side-panel-test.js
@@ -19,7 +19,8 @@ const interaction = {
   advisorId: '11',
   clientContactId: '299',
   projectId: '14',
-  used: false
+  used: false,
+  paymentRequired: true
 };
 
 QUnit.module("Scheduled Interactions Side Panel", {
@@ -98,7 +99,8 @@ QUnit.module("Scheduled Interactions Side Panel", {
           "advisor_phone_country_code": interaction.advisorPhoneCountryCode,
           "advisor_phone_number": interaction.advisorPhoneNumber,
           "speak_phone_number": interaction.speakPhoneNumber,
-          "speak_code": interaction.speakCode
+          "speak_code": interaction.speakCode,
+          "payment_required": interaction.paymentRequired
         }
       ],
       "checklist_items": []
@@ -127,7 +129,8 @@ test("Cancel interaction returns to dashboard and removes interaction from the w
         "dial_in_number": "1234567",
         "primary_contact_id": 1,
         "speak": true,
-        "client_access_number_country": "US"
+        "client_access_number_country": "US",
+        "payment_required": true
       }
     ]
   }});
@@ -230,6 +233,7 @@ test("Reschedule Interaction unschedules the call and transitions to the to sche
     "interaction": {
       "actioned": interaction.actioned,
       "client_access_number_country": interaction.clientAccessNumberCountry,
+      "payment_required": interaction.paymentRequired,
       "additional_contact_details": interaction.additionalContactDetails,
       "requested_at": interaction.requestedAt,
       "scheduled_call_time": null,
@@ -239,10 +243,10 @@ test("Reschedule Interaction unschedules the call and transitions to the to sche
       "advisor_phone_number": interaction.advisorPhoneNumber,
       "speak_phone_number": interaction.speakPhoneNumber,
       "speak_code": interaction.speakCode,
+      "used": interaction.used,
       "advisor_id": interaction.advisorId,
       "client_contact_id": interaction.clientContactId,
-      "project_id": interaction.projectId,
-      "used": interaction.used
+      "project_id": interaction.projectId
     }
   }});
 
@@ -324,7 +328,7 @@ const interactionCompletion = {
   speakExplanation: 'Client was grumpy.'
 };
 
-test("Complete Interaction completes the call and closes the side panel", function(assert) {
+test("Complete Interaction and Charge Client completes the call", function(assert) {
   const successMessage = 'The interaction has been completed.';
 
   var handler = defineFixture('POST', '/interaction_completions', {
@@ -374,39 +378,75 @@ test("Complete Interaction completes the call and closes the side panel", functi
   andThen(function() {
     assert.equal(handler.called, true);
     assert.equal($('.messenger-message-inner:first').text().trim(), successMessage);
-    assert.equal(find('.scheduled-interactions article').length, 0);
-    assert.equal(currentURL(), '/dashboard');
   });
 });
 
-test("Complete Interaction shows error message in case of failure", function(assert) {
-  var handler = defineFixture('POST', '/interaction_completions', {
-    status: 500,
-
-    request: {
-      "interaction_completion": {
-        "duration": interactionCompletion.duration,
-        "quality": interactionCompletion.quality,
-        "interaction_type": interactionCompletion.interactionType,
-        "interaction_id": interactionCompletion.interactionId,
-        "speak_quality": interactionCompletion.speakQuality,
-        "speak_explanation": interactionCompletion.speakExplanation
-      }
+test("Complete Interaction and Don't Pay Advisor/Change updates the interaction", function(assert) {
+  let dontPayAdvisorRequest = defineFixture('PUT', `/interactions/${interaction.id}`, { request: {
+    "interaction": {
+      "actioned": interaction.actioned,
+      "client_access_number_country": interaction.clientAccessNumberCountry,
+      "additional_contact_details": interaction.additionalContactDetails,
+      "payment_required": false,
+      "requested_at": interaction.requestedAt,
+      "scheduled_call_time": interaction.scheduledCallTime,
+      "speak": interaction.speak,
+      "interaction_type": interaction.interactionType,
+      "advisor_phone_country_code": interaction.advisorPhoneCountryCode,
+      "advisor_phone_number": interaction.advisorPhoneNumber,
+      "speak_phone_number": interaction.speakPhoneNumber,
+      "speak_code": interaction.speakCode,
+      "used": interaction.used,
+      "advisor_id": interaction.advisorId,
+      "client_contact_id": interaction.clientContactId,
+      "project_id": interaction.projectId
     }
-  });
+  }});
+
 
   visit('/dashboard');
 
   click('.scheduled-interactions article:first');
   click('button:contains("Complete Interaction")');
-  fillIn('input[name=duration]', '20');
-  select('select[name=quality] ', 'Bad');
-  select('select[name=speakQuality] ', 'Other issue');
-  fillIn('input[name=speakExplanation]', interactionCompletion.speakExplanation);
-  click('button:contains("Charge Client")');
+
+  let buttonText = "Don't Pay";
+  click(`button:contains(${buttonText})`);
 
   andThen(function() {
-    const message = 'There has been an error completing the interaction.';
-    assert.equal(message, $('.messenger .messenger-message-inner').first().text().trim());
+    const message = 'The advisor will not be paid.';
+    assert.equal(
+      message, $('.advisor-payment .actions span').first().text().trim(),
+      "it updates the content to say the advisor will not be paid"
+    );
+  });
+
+  let changeAdvisorPaymentRequest = defineFixture('PUT', `/interactions/${interaction.id}`, { request: {
+    "interaction": {
+      "actioned": interaction.actioned,
+      "client_access_number_country": interaction.clientAccessNumberCountry,
+      "additional_contact_details": interaction.additionalContactDetails,
+      "payment_required": true,
+      "requested_at": interaction.requestedAt,
+      "scheduled_call_time": interaction.scheduledCallTime,
+      "speak": interaction.speak,
+      "interaction_type": interaction.interactionType,
+      "advisor_phone_country_code": interaction.advisorPhoneCountryCode,
+      "advisor_phone_number": interaction.advisorPhoneNumber,
+      "speak_phone_number": interaction.speakPhoneNumber,
+      "speak_code": interaction.speakCode,
+      "used": interaction.used,
+      "advisor_id": interaction.advisorId,
+      "client_contact_id": interaction.clientContactId,
+      "project_id": interaction.projectId
+    }
+  }});
+
+  click("button:contains('Change')");
+
+  andThen(function() {
+    assert.equal(
+      $(`button:contains(${buttonText})`).length, 1,
+      "it displays the Don't Pay button again after clicking 'Change'"
+    );
   });
 });